Deleted Changes and modified according to code review
